// src/generated/prisma/enums.ts
export const AiSystemStatus = {
  DRAFT: 'DRAFT',
  IN_REVIEW: 'IN_REVIEW',
  ACTION_REQUIRED: 'ACTION_REQUIRED',
  APPROVED: 'APPROVED',
  CONDITIONALLY_APPROVED: 'CONDITIONALLY_APPROVED',
  REJECTED: 'REJECTED',
  SUSPENDED: 'SUSPENDED'
} as const

export type AiSystemStatus = (typeof AiSystemStatus)[keyof typeof AiSystemStatus]